Move PaywallComponent Array/Dict extensions to own file
